requirements.txt
requests>=2.25.0
pillow>=8.0.0 
